// src/types/actors/actrosDetals.ts
export type ActrosDetals = {
	adult: boolean;
	also_known_as: string[];
	biography: string;
	birthday: string;
	deathday: boolean | string;
	gender: 0 | 1 | 2 | 3;
	homepage: boolean;
	id: number;
	imdb_id: string;
	known_for_department: string;
	name: string;
	place_of_birth: string;
	popularity: number;
	profile_path: string;
};